requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1

# Зависимости для тестирования
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
